remove unnecessary code

+++ src/main/scala/com/github/arzt/tensor/convert/ByteToBooleanConverter.scala
@@ -1,7 +1,7 @@
 package com.github.arzt.tensor.convert
 
-import com.github.arzt.tensor.convert.Converter.readInt
-import com.github.arzt.tensor.convert.Converter.writeInt
+import com.github.arzt.tensor.convert.Converter.readLong
+import com.github.arzt.tensor.convert.Converter.writeLong
 
 class ByteToBooleanConverter extends Converter[Byte, Boolean] {
 
@@ -10,8 +10,8 @@ class ByteToBooleanConverter extends Converter[Byte, Boolean] {
   override val n = 8
 
   override def read(s: Byte, i: Int): Boolean =
-    (readInt(1, s, n - i - 1) & 1) == 1
+    (readLong(1, s, n - i - 1) & 1) == 1
 
   override def write(s: Byte, i: Int, t: Boolean): Byte =
-    writeInt(1, s, n - i - 1, if (t) 1 else 0).toByte
+    writeLong(1, s, n - i - 1, if (t) 1 else 0).toByte
 }
